pyproject.toml
[tool.poetry]
name = "ocfkube"
version = "0.1.0"
description = "OCF Kubernetes configuration"
authors = [
    "OCF Staff <root@ocf.berkeley.edu>",
    "Nikhil Jha <njha@ocf.berkeley.edu>",
    "Frank Dai <fydai@ocf.berkeley.edu>",
]

[tool.poetry.dependencies]
python = "^3.7"
kubernetes = "^17.17.0"
PyYAML = "^5.4.1"
tomlkit = "^0.7.2"

[tool.poetry.scripts]
argocd-build = "ocfkube:argocd_build"

[tool.poetry.dev-dependencies]
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
